requirements.txt
openai
pytest
requests
flask
gitpython
